pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"
pattern = "default-unprefixed"

[tool.poetry.dependencies]
python = "^3.8"
httpx = { version = ">=0.15.4,<0.27.2", extras = ["http2"] }
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"
PyJWT = "^2.0.0"
protobuf = "^4.0.0"

[tool.poetry]
name = "neptune-api"
version = "0.1.0"
description = "A client library for accessing Neptune API"
authors = ["neptune.ai <contact@neptune.ai>"]
repository = "https://github.com/neptune-ai/neptune-api"
readme = "README.md"
license = "Apache License 2.0"
packages = [
    { include = "neptune_api", from = "src" },
    { include = "neptune_retrieval_api", from = "src" },
]
include = ["CHANGELOG.md", "src/neptune_api/py.typed"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = [
    "MLOps",
    "ML Experiment Tracking",
    "ML Model Registry",
    "ML Model Store",
    "ML Metadata Store",
]

[tool.poetry.urls]
"Tracker" = "https://github.com/neptune-ai/neptune-api/issues"
"Documentation" = "https://docs.neptune.ai/"

[tool.black]
line-length = 120
target_version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$,\_pb2\.py$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
force_grid_wrap = 2
extend_skip = ["src/neptune_api/proto", "src/neptune_retrieval_api/proto"]

[tool.ruff]
line-length = 120
exclude = ["src/neptune_api/proto", "src/neptune_retrieval_api/proto"]

[tool.ruff.lint]
select = ["F", "UP"]

[tool.mypy]
files = 'src/neptune_api'
mypy_path = "stubs"
install_types = "True"
non_interactive = "True"
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
show_error_codes = "True"
warn_unused_ignores = "True"
